+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Go Library
 
+## [v1.7.0 - 2023-04-24]
+### New
+- add [Hashicorp Vault](https://developer.hashicorp.com/vault) KV2 and Logical API functions
+
 ## [v1.6.0 - 2023-04-09]
 ### New
 - maillib: add Imap functions